+++ trunk/moses-cmd/src/ThreadPool.cpp
@@ -55,7 +55,7 @@ void Moses::ThreadPool::Execute()
         m_threadAvailable.notify_all();
     } while (!m_stopped);
 #if defined(BOOST_HAS_PTHREADS)
-    TRACE_ERR("Thread " << (int)pthread_self() << " exiting" << endl);
+    TRACE_ERR("Thread " << pthread_self() << " exiting" << endl);
 #endif
 }
